fix(checker): using-declaration missing dispose member reports TS2741, not TS2850
When a sync `using` initializer is object-like and merely lacks
`[Symbol.dispose]`, tsc's checkTypeAssignableTo(initType, Disposable,
headMessage=TS2850) drops the headMessage frame and reports the relation's
own missing-property diagnostic (TS2741) directly, since the "must be an
object with a [Symbol.dispose]() method..." wording doesn't apply when the
initializer already is such an object. tsz reported TS2850 uniformly. A
present-but-incompatible member (wrong type/arity) and a wholly non-object
initializer keep TS2850, oracle-verified against typescript@7.0.2.

Extracted the using-declaration disposable-check section from core.rs into
its own module (using_declaration_disposable.rs) to stay under the 2000-line
checker-file cap after the new logic.

Closes #16872.
